Add cards module, change board and player intrctn
Tracking the Risk cards is now handled by the Cards class, found in
the cards module. This class provides easy access to all relevant
information. The interaction between board a player is now changed:
the player contains references to all relevant objects: the Board,
his Mission and his Cards. This means that the Game does not need
to pass self to each call. Also, the Player is split up into a base
player and Attack/Reinforce/Fortify Mixins, such that it is easy
to combine these.